create new component with searchFormContainer
