+++ src/components/search/searchFormContainer.js
@@ -1,0 +1,79 @@
+import React, { useState } from "react";
+import { useJsApiLoader } from "@react-google-maps/api";
+import Search from "./searchInput";
+import RangeSlider from "./timeRangeSlider";
+import SearchButton from "./searchButton";
+import Categories from "./categories";
+import CategoryA from "./../../images/CategoryA.png"
+import CategoryB from "./../../images/CategoryB.png"
+import CategoryC from "./../../images/CategoryC.png"
+import CategoryD from "./../../images/CategoryD.png"
+import "./searchFormContainer.css";
+
+const libraries = ["places"];
+
+function SearchFormContainer({ routeResponse, setRouteResponse }) {
+
+    const { isLoaded } = useJsApiLoader({
+        id: "google-map-script",
+        googleMapsApiKey: process.env.REACT_APP_GOOGLE_MAPS_API_KEY,
+        libraries,
+    });
+
+    const mapRef = React.useRef();
+
+    const panTo = React.useCallback(({ lat, lng }) => {
+        mapRef.current.panTo({ lat, lng });
+        mapRef.current.setZoom(16);
+    }, []);
+
+    const [to, setTo] = useState("");
+    const [from, setFrom] = useState("");
+    const [time, setTime] = useState("");
+    const [selectedCategory, setSelectedCategory] = useState();
+
+    return (
+        <div>
+            <div className="searchFormContainer">
+                <form className="searchForm">
+                    <div class="form-row inputContainer">
+                        <div class="col-md-5 whiteBg">
+                            {isLoaded && <Search className="TiconFrom" onChange={setFrom} panTo={panTo.bind(this)} placeholder={'FROM...'} />}
+                            {/* <CurrentLocation panTo={panTo} /> */}
+                        </div>
+                        <div class="col-md-5 whiteBg">
+                            {isLoaded && <Search onChange={setTo} panTo={panTo.bind(this)} placeholder={'TO...'} />}
+                        </div>
+                    </div>
+
+                    <p>Tell us about your mood today...</p>
+                    <div className="categoryImages">
+                        <Categories className={'Berliner_Restaurant'} src={CategoryA} setSelectedCategory={setSelectedCategory} />
+                        <Categories className={'Eiscafe'} src={CategoryB} setSelectedCategory={setSelectedCategory} />
+                        <Categories className={'categoryC'} src={CategoryC} setSelectedCategory={setSelectedCategory} />
+                        <Categories className={'categoryD'} src={CategoryD} setSelectedCategory={setSelectedCategory} />
+                        <Categories className={'categoryE'} src={CategoryA} setSelectedCategory={setSelectedCategory} />
+                    </div>
+                    <div>
+                        <p class="whiteBg">How much time do you have? {time}' min</p>
+                        <div class="form-group slideRanger">
+                            <RangeSlider onChange={setTime} />
+                        </div>
+                    </div>
+
+                    <div class="searchDirectionButton">
+                        {isLoaded && <SearchButton
+                            to={to}
+                            from={from}
+                            routeResponse={routeResponse}
+                            setRouteResponse={setRouteResponse}
+                            selectedCategory={selectedCategory}
+                        />}
+                    </div>
+                </form>
+            </div>
+        </div>
+    );
+}
+
+export default SearchFormContainer;
